Generar practicas

+++ server/controllers/ejercicioController.js
@@ -1,5 +1,8 @@
 const { validationResult } = require('express-validator')
 const Ejercicio = require('../models/Ejercicio')
+const Plantilla = require('../models/Plantilla')
+const Calificacion = require('../models/Calificacion')
+const Practica = require('../models/Practica')
 const { asignaturasCoordinador } = require('../utils/coordinador')
 const { asignaturasDocente } = require('../utils/docente')
 
@@ -69,6 +72,7 @@ exports.buscarEjercicios = async (req, res) => {
   try {
     // buscar en la db
     const ejercicios = await Ejercicio.find()
+      .sort({ _id: -1 })
       .populate({ path: 'tema', select: 'nombre' })
       .populate({ path: 'asignatura', select: 'nombre' })
       .populate({ path: 'docente', select: '-contrasena' })
@@ -148,7 +152,7 @@ exports.modificarEjercicio = async (req, res) => {
 exports.eliminarEjercicio = async (req, res) => {
   try {
     // Revisar si existe por el id enviado
-    const ejercicioEncontrada = await Ejercicio.findById(req.params.id)
+    let ejercicioEncontrada = await Ejercicio.findById(req.params.id)
     if (!ejercicioEncontrada) {
       res.status(404).json({ msg: 'Ejercicio a eliminar no encontrada' })
       return
@@ -175,6 +179,24 @@ exports.eliminarEjercicio = async (req, res) => {
       }
     }
 
+    // En caso de que el ejercicio este usado en una practica archivarlo
+    const usado = await Practica.find({ ejercicios: ejercicioEncontrada._id })
+
+    if (usado.length > 0) {
+      ejercicioEncontrada = await Ejercicio.findByIdAndUpdate(
+        { _id: req.params.id },
+        { $set: { archivado: true } },
+        { new: true }
+      )
+
+      res.status(200).json({
+        msg: 'Ejercicio ya utilizado ha sido archivado',
+        data: ejercicioEncontrada,
+      })
+
+      return
+    }
+
     // Eliminar en la db
     await Ejercicio.findOneAndRemove({ _id: req.params.id })
     res.status(200).json({ msg: 'Ejercicio eliminada con exito' })
@@ -188,6 +210,7 @@ exports.buscarEjercicioAsignatura = async (req, res) => {
   try {
     // buscar en la db
     const ejercicios = await Ejercicio.find({ asignatura: req.params.id })
+      .sort({ _id: -1 })
       .populate({ path: 'tema', select: 'nombre' })
       .populate({ path: 'asignatura', select: 'nombre' })
       .populate({ path: 'docente', select: '-contrasena' })
@@ -200,10 +223,102 @@ exports.buscarEjercicioAsignatura = async (req, res) => {
       return
     }
 
+    // CAMBIAR POR LOOKUP
+    const nuevoEjercicio = []
+    await Promise.all(
+      ejercicios.map(async (ejercicio) => {
+        const calificacionEjer = await Calificacion.find({
+          ejercicio: ejercicio._id,
+        })
+
+        const ejer = {
+          referencia: ejercicio.referencia,
+          _id: ejercicio._id,
+          titulo: ejercicio.titulo,
+          descripcion: ejercicio.descripcion,
+          dificultad: ejercicio.dificultad,
+          tema: ejercicio.tema,
+          asignatura: ejercicio.asignatura,
+          docente: ejercicio.docente,
+          evaluacion: ejercicio.evaluacion,
+          archivado: ejercicio.archivado,
+          calificacion: calificacionEjer,
+        }
+        // console.log(ejer.calificacion)
+        // // ejercicio.calificacion = calificacionEjer
+        nuevoEjercicio.push(ejer)
+      })
+    )
+
+    nuevoEjercicio.sort((a, b) => {
+      return a.titulo.localeCompare(b.titulo)
+    })
+
     // caso contrario retornar la lista
     res.status(200).json({
       msg: 'Busqueda realizada con exito',
-      data: ejercicios,
+      data: nuevoEjercicio,
+    })
+  } catch (error) {
+    res.status(500).json({ msg: 'hubo un error en el servidor' })
+  }
+}
+
+// const numUsado = async item => {
+//   return Practica.find({ tema: plantilla.temas })
+// }
+
+// buscarEjercicioPlantilla Busca todas las ejercicios en la base de datos del tema de una plantilla
+exports.buscarEjercicioPlantilla = async (req, res) => {
+  try {
+    // tomar el tema
+    const plantilla = await Plantilla.findById(req.params.id)
+    // si no hay datos retornar 404 not found
+    if (!plantilla) {
+      res.status(404).json({ msg: 'No se encontro la plantilla a ingrear' })
+      return
+    }
+
+    // buscar en la db
+    const ejercicios = await Ejercicio.find({
+      tema: plantilla.temas,
+      evaluacion: 0,
+      archivado: false,
+    }).populate({ path: 'tema', select: 'nombre' })
+    // si no hay datos retornar 404 not found
+    if (!ejercicios) {
+      res.status(404).json({ msg: 'No se encontraron ejercicios' })
+      return
+    }
+
+    // CAMBIAR POR LOOKUP
+    const nuevoEjercicio = []
+    await Promise.all(
+      ejercicios.map(async (ejercicio) => {
+        const num = await Practica.find({ ejercicios: ejercicio._id })
+
+        const ejer = {
+          referencia: ejercicio.referencia,
+          _id: ejercicio._id,
+          titulo: ejercicio.titulo,
+          descripcion: ejercicio.descripcion,
+          dificultad: ejercicio.dificultad,
+          tema: ejercicio.tema,
+          asignatura: ejercicio.asignatura,
+          docente: ejercicio.docente,
+          archivado: ejercicio.archivado,
+          usado: num.length,
+        }
+        // console.log(ejer.calificacion)
+        // // ejercicio.calificacion = calificacionEjer
+        nuevoEjercicio.push(ejer)
+      })
+    )
+
+    // caso contrario retornar la lista
+    res.status(200).json({
+      msg: 'Busqueda realizada con exito',
+      data: nuevoEjercicio,
     })
   } catch (error) {
     res.status(500).json({ msg: 'hubo un error en el servidor' })
